Refactoring hdf5 and mat input nodes - Integrating code from HDF5FileNode and MatlabFileNode
